+++ src/Styles/Yeseul/base.scss
@@ -2,42 +2,24 @@
 
 /* layout */
 .align-item-center {
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 .space-between {
-    display: flex;
-    justify-content: space-between;
+  display: flex;
+  justify-content: space-between;
 }
 
 .align-right {
-    margin-left: auto;
+  margin-left: auto;
 }
 
 /* style */
 
-button{
-    width: 40px;
-    height: 40px;
-    font-weight: 600;
-    color: $btn--blue;
-  }
-
-.user-name {
-    margin-right: 4px;
-    color: $main-font;
-    font-weight: 800;
-    font-size: 14px;
-    cursor: pointer;
-}
-
-.user-name:hover {
-    text-decoration: underline;
-}
-
-.detail-info {
-    color: $font--grey;
-    font-size: 12px;
-    line-height: 14px;
+button {
+  width: 40px;
+  height: 40px;
+  font-weight: 600;
+  color: $btn--blue;
 }
