--- src/pages/Yeseul/Main/Components/Recommendation/Recommendation.scss
@@ -1,25 +1,0 @@
-@import '../../../../../Styles/Yeseul/variable.scss';
-
-.recommendation {
-  $self: &;
-  padding: 8px 16px;
-
-  #{$self}__header {
-    padding: 8px 0;
-
-    #{$self}__title {
-      color: $font--grey;
-      font-size: 14px;
-    }
-
-    a {
-      color: $main-font;
-      font-size: 12px;
-    }
-  }
-}
-  
-.follow {
-    margin-right: -8px;
-    width: auto;
-}
