+++ src/pages/Yeseul/Main/Components/SectionRight/SectionRight.scss
@@ -1,16 +1,15 @@
 @import '../../../../../Styles/Yeseul/variable.scss';
 
-.story {
-  $self: &;
-  margin: 16px 0;
+.sectionRight {
+  margin-top: 16px;
   padding: 8px 16px;
   height: 220px;
   overflow: auto;
 
-  &__header {
+  .title {
     padding: 8px 0;
 
-    #{$self}__title {
+    span {
       color: $font--grey;
       font-size: 14px;
     }
@@ -20,4 +19,9 @@
       font-size: 12px;
     }
   }
+}
+
+.follow {
+  margin-right: -8px;
+  width: auto;
 }
