+++ src/pages/Yeseul/Main/Components/UserList/UserList.scss
@@ -1,0 +1,5 @@
+.userList {
+  li {
+    padding: 8px 0;
+  }
+}
