--- src/pages/Yeseul/Main/Components/UsrList/UserList.scss
@@ -1,3 +1,0 @@
-li[role="user-list"] {
-  padding: 8px 0;
-}
